+++ db/migrate/001_add_project_type_id_to_projects.rb
@@ -19,7 +19,7 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-class AddProjectTypeIdToProjects < ActiveRecord::Migration[4.2]
+class AddProjectTypeIdToProjects < ActiveRecord::Migration[4.2] #:nodoc:
   def change
     add_column :projects, :project_type_id, :integer, null: true
   end