+++ test/functional/custom_fields_controller_test.rb
@@ -39,9 +39,9 @@ module ProjectTypes
     end
 
     test 'should create new issue custom field' do
-      project_type2 = create_project_type(name: 'ProjectType2').relative_ids = [1]  
-      project_type3 = create_project_type(name: 'ProjectType3').relative_ids = [2]
-      project_type4 = create_project_type(name: 'ProjectType4').relative_ids = [3]
+      create_project_type(name: 'ProjectType2').relative_ids = [1]
+      create_project_type(name: 'ProjectType3').relative_ids = [2]
+      create_project_type(name: 'ProjectType4').relative_ids = [3]
 
       get new_custom_field_path, params: { type: 'IssueCustomField' }
       assert_response :success