+++ test/unit/patches/project_test_patch.rb
@@ -20,27 +20,27 @@
 
 class ProjectTest
   define_method('test_safe_attributes_should_include_only_custom_fields_visible_to_user') do
-    cf1 = ProjectCustomField.create!(:name => 'Visible field',
-                                   :field_format => 'string',
-                                   :visible => false, :role_ids => [1])
-    cf2 = ProjectCustomField.create!(:name => 'Non visible field',
-                                   :field_format => 'string',
-                                   :visible => false, :role_ids => [3])
+    cf1 = ProjectCustomField.create!(name: 'Visible field',
+                                     field_format: 'string',
+                                     visible: false, role_ids: [1])
+    cf2 = ProjectCustomField.create!(name: 'Non visible field',
+                                     field_format: 'string',
+                                     visible: false, role_ids: [3])
     user = User.find(2)
     project = Project.find(1)
     project.project_custom_field_ids = [cf1.id, cf2.id]
-    project.save 
+    project.save
 
-    project.send :safe_attributes=, {'custom_field_values' => {
-                                       cf1.id.to_s => 'value1', cf2.id.to_s => 'value2'
-                                     }}, user
+    project.send :safe_attributes=, { 'custom_field_values' => {
+      cf1.id.to_s => 'value1', cf2.id.to_s => 'value2'
+    } }, user
     assert_equal 'value1', project.custom_field_value(cf1)
     assert_nil project.custom_field_value(cf2)
 
-    project.send :safe_attributes=, {'custom_fields' => [
-                                      {'id' => cf1.id.to_s, 'value' => 'valuea'},
-                                      {'id' => cf2.id.to_s, 'value' => 'valueb'}
-                                    ]}, user
+    project.send :safe_attributes=, { 'custom_fields' => [
+      { 'id' => cf1.id.to_s, 'value' => 'valuea' },
+      { 'id' => cf2.id.to_s, 'value' => 'valueb' }
+    ] }, user
     assert_equal 'valuea', project.custom_field_value(cf1)
     assert_nil project.custom_field_value(cf2)
   end