+++ app/controllers/project_types_controller.rb
@@ -33,7 +33,7 @@ class ProjectTypesController < ApplicationController #:nodoc:
   def index
     @status = params[:status] || 1
 
-    scope = ProjectType.masters.status(@status).sorted
+    scope = ProjectType.masters.includes(:subordinates).status(@status).sorted
     scope = scope.like(params[:name]) if params[:name].present?
 
     @project_count = scope.count